check that backup source is not None added and backup task renamed (nc-195)

+++ nodeconductor/backup/tasks.py
@@ -9,19 +9,27 @@ logger = logging.getLogger(__name__)
 
 
 @shared_task
-def backup_task(backup_uuid):
+def process_backup_task(backup_uuid):
     try:
         backup = models.Backup.objects.get(uuid=backup_uuid)
-        backup.backup_source.get_backup_strategy().backup()
+        source = backup.backup_source
+        if source is not None:
+            backup.backup_source.get_backup_strategy().backup()
+        else:
+            logger.exception('Process backup task was called for backup with no source. Backup uuid: %s', backup_uuid)
     except models.Backup.DoesNotExist:
-        logger.exception('Backup task was called for backed with uuid %s which does not exist', backup_uuid)
+        logger.exception('Process backup task was called for backed with uuid %s which does not exist', backup_uuid)
 
 
 @shared_task
 def restoration_task(backup_uuid, replace_original=False):
     try:
         backup = models.Backup.objects.get(uuid=backup_uuid)
-        backup.backup_source.get_backup_strategy().restore(replace_original)
+        source = backup.backup_source
+        if source is not None:
+            backup.backup_source.get_backup_strategy().restore(replace_original)
+        else:
+            logger.exception('Restoration task was called for backup with no source. Backup uuid: %s', backup_uuid)
     except models.Backup.DoesNotExist:
         logger.exception('Restoration task was called for backed with uuid %s which does not exist', backup_uuid)
 
@@ -30,6 +38,10 @@ def restoration_task(backup_uuid, replace_original=False):
 def deletion_task(backup_uuid):
     try:
         backup = models.Backup.objects.get(uuid=backup_uuid)
-        backup.backup_source.get_backup_strategy().delete()
+        source = backup.backup_source
+        if source is not None:
+            backup.backup_source.get_backup_strategy().delete()
+        else:
+            logger.exception('Restoration task was called for backup with no source. Backup uuid: %s', backup_uuid)
     except models.Backup.DoesNotExist:
         logger.exception('Deletion task was called for backed with uuid %s which does not exist', backup_uuid)
